+++ src/object.rs
@@ -3,17 +3,17 @@ mod expo_push_error_receipt;
 mod expo_push_error_ticket;
 mod expo_push_message;
 mod expo_push_receipt;
-mod expo_push_receipt_id;
 mod expo_push_success_receipt;
 mod expo_push_success_ticket;
 mod expo_push_ticket;
+mod get_push_notification_receipts_request;
 
 pub use self::details::*;
 pub use self::expo_push_error_receipt::*;
 pub use self::expo_push_error_ticket::*;
 pub use self::expo_push_message::*;
 pub use self::expo_push_receipt::*;
-pub use self::expo_push_receipt_id::*;
 pub use self::expo_push_success_receipt::*;
 pub use self::expo_push_success_ticket::*;
 pub use self::expo_push_ticket::*;
+pub use self::get_push_notification_receipts_request::*;
